component support for Cassandra connections and queries
Based on the discussion in #51, add support for component.
Two variants of the same component are provided to establish
sessions with clusters.

- `cassandra-registry` connects to a single cluster
- `map->CassandraMultiClusterRegistry` reaches out to several

The idea is that a downstream component should depend on one of
these registries and then use query-functions to get a map of
keyword to query.

For convenience purposes, a QueryRegistry component is also
supplied, which takes care of building an appropriate query
function, which can be called either directly or through query! on the
component, matching standard behavior of storage components.